Increase GHA timeout for tests.
ABQ recently changed to use a timeout when an underlying native process
fails instead of exiting immediately, in order to capture all the child
output. We expect this to be mitigated in the future, but for now we'll
increase the timeouts a bit to account for a few tests we have that
are testing failure scenarios.