Register the default RSpec formatter when none are configured. (#110)
* Register the default RSpec formatter when none are configured.

RSpec typically adds its default "progress" formatter when no formatters
have been manually registered. Registering RSpec::Abq::Formatter causes
RSpec to skip this.

We recreate this functionality by adding the default formatter when no
formatters have been registered yet. This doesn't result in duplication
of the same formatter. Users can still manually configure their own
formatters in RSpec configuration.

* Bump version to 1.1.4.

* Bump rack to 2.2.6.4+ for security patches.

* Update test expectations for multiple versions of ABQ.

* Increase GHA timeout for tests.

ABQ recently changed to use a timeout when an underlying native process
fails instead of exiting immediately, in order to capture all the child
output. We expect this to be mitigated in the future, but for now we'll
increase the timeouts a bit to account for a few tests we have that
are testing failure scenarios.